change ex and terminal page and alose readme
