+++ README.md
@@ -1,9 +1,72 @@
-# [As Much何とか](https://www.asmuch.xyz) [![CircleCI](https://circleci.com/gh/asmuch/asmuch.svg?style=shield)](https://circleci.com/gh/asmuch/asmuch)
+# Welcome to As Much 何とか 👋
+![Version](https://img.shields.io/badge/version-1.3.0-blue.svg?cacheSeconds=2592000)
+[![CircleCI](https://circleci.com/gh/asmuch/asmuch.svg?style=shield)](https://circleci.com/gh/asmuch/asmuch)
+[![Documentation](https://img.shields.io/badge/documentation-yes-brightgreen.svg)](https://github.com/asmuch/asmuch/wiki)
+[![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://github.com/asmuch/asmuch/blob/master/LICENSE.md)
+[![Twitter: faridh212](https://img.shields.io/twitter/follow/faridh212.svg?style=social)](https://twitter.com/faridh212)
 
-Community for people who are willing to help eachother to learn code.
+[![Twitter: natural_clar](https://img.shields.io/twitter/follow/natural_clar.svg?style=social)](https://twitter.com/faridh212)
+> Official web site of As Much group, group of dev that want to share and exchange knowledge
 
+### 🏠 [Homepage](https://www.asmuch.xyz/)
+
+## Install
+
+```sh
+yarn install
+```
+
+## Usage
+
+```sh
+yarn start
+```
+
+## Run tests
+
+```sh
+yarn test
+```
+
+## Author
+
+👤 **Farid Hammouten** - *Fullstack Javascript*
+
+* Twitter: [@faridh212](https://twitter.com/faridh212)
+* Github: [@farid212](https://github.com/farid212)
 
 ## Member
-- [Farid212](https://github.com/Farid212) - **Farid Hammouten**
-- [Naturalclar](https://github.com/Naturalclar) - **Jesse Katsumata**
-- [Phantrandai](https://github.com/phantrandai) - **Dai Phan Tran**
+
+👤 **Jesse Katsumata** - *Fullstack Javascript*
+
+* Twitter: [@natural_clar](https://twitter.com/natural_clar)
+* Github: [@NaturalClar](https://github.com/NaturalClar)
+
+👤 **Oussama Ladhari** - *Fullstack Javascript*
+
+* Github: [@oladhari](https://github.com/oladhari)
+* Facebook: [@oussama.ladhari.9](https://www.facebook.com/oussama.ladhari.9)
+
+👤 **Dai Phan Tran** -  *Web Designer*
+
+* Github: [@Phantrandai](https://github.com/phantrandai)
+
+## 🤝 Contributing
+
+Contributions, issues and feature requests are welcome!
+
+Feel free to check [issues page](https://github.com/asmuch/asmuch/issues).
+
+## Show your support
+
+Give a ⭐️ if this project helped you!
+
+
+## 📝 License
+
+Copyright © 2019 [Farid Hammouten](https://github.com/farid212).
+
+This project is [MIT](https://github.com/asmuch/asmuch/blob/master/LICENSE.md) licensed.
+
+***
+_This README was generated with ❤️ by [readme-md-generator](https://github.com/kefranabg/readme-md-generator)_
