+++ src/routes/index.jsx
@@ -1,10 +1,10 @@
-import LandingPage from "views/LandingPage/LandingPage";
-import TerminalPage from "views/TerminalPage/TerminalPage";
-import FaProfilePage from "views/ProfilePage/FaProfilePage";
+import LandingPage from "../views/LandingPage/LandingPage";
+import ExoPage from "../views/ExoPage/ExoPage.jsx";
+import FaProfilePage from "../views/ProfilePage/FaProfilePage";
 
 var indexRoutes = [
   { path: "/", name: "LandingPage", component: LandingPage },
-  { path: "/terminal", name: "TerminalPage", component: TerminalPage },
+  { path: "/Exo", name: "ExoPage", component: ExoPage },
   { path: "/farid", name: "ProfilePage", component: FaProfilePage }
 ];
