+++ src/views/LandingPage/Sections/TerminalSection/partials/fileSys.js
